feat: Implement Google Sign-In only for web
- Google authentication is now restricted to the web application.
- Android and iOS authentication are not supported.
- Detailed explanation available in the issue assigned to this feature.
