+++ ThriftUni/package.json
@@ -5,8 +5,8 @@
   "version": "1.0.0",
   "scripts": {
     "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
     "web": "expo start --web",
     "test": "jest --watchAll"
   },
@@ -17,7 +17,8 @@
     "@expo/vector-icons": "^14.0.2",
     "@react-navigation/native": "^7.0.14",
     "diff": "^7.0.0",
-    "expo": "^52.0.32",
+    "expo": "~52.0.33",
+    "expo-auth-session": "~6.0.3",
     "expo-font": "~13.0.3",
     "expo-linking": "~7.0.5",
     "expo-router": "^4.0.17",
@@ -25,7 +26,7 @@
     "expo-status-bar": "~2.0.1",
     "expo-system-ui": "~4.0.7",
     "expo-web-browser": "~14.0.2",
-    "firebase": "^11.2.0",
+    "firebase": "^11.3.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.6",
